requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary  # Для PostgreSQL
pydantic[email]
python-jose[cryptography] # Для JWT
passlib[bcrypt] # Для хеширования паролей
python-dotenv # Для загрузки .env файла
pydantic_settings
cryptography